pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "apscheduler>=3.6.3",
    "base58>=2.1.1",
    "ccxt>=4.4.30",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "python-telegram-bot==13.7",
    "pytz>=2025.1",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "solana>=0.25.1",
    "solana-sdk>=0.25.6",
    "telegram>=0.0.1",
]
